6.10 继续优化 最大程度精简这个项目

+++ music-server/src/main/java/com/example/yin/controller/SingerController.java
@@ -40,7 +40,7 @@ public class SingerController {
 
     // 根据歌手性别查找歌手
     @GetMapping("/singer/sex/detail")
-    public R singerOfSex(@RequestParam int sex) {
+    public R singerOfSex(@RequestParam Byte sex) {
         return singerService.singerOfSex(sex);
     }
